+++ helper/jwtDecode.ts
@@ -1,8 +1,13 @@
 import { jwtDecode } from "jwt-decode";
 
 export const jwtDecoder = (token:any) => {
-  if (token) {
-    const tokenDecoded = jwtDecode(token || null);
-    return tokenDecoded;
+  try {
+    if (token) {
+      const tokenDecoded = jwtDecode(token);
+      return tokenDecoded;
+    }
+  } catch (error) {
+    console.error('JWT Decoding Error:', error);
+    // Handle the error appropriately, such as returning null or throwing a custom error.
   }
 };
