+++ tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "redux/store.ts", "helper/axiosInstance.ts", "helper/handleRequest.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "redux/store.ts", "helper/axiosInstance.ts", "helper/handleRequest.ts"],
   "exclude": ["node_modules"]
 }
